requirements.txt
anthropic>=0.40.0
streamlit>=1.40.0
python-dotenv>=1.0.0
pydantic>=2.0.0
